requirements.txt
asgiref==3.2.3
Django==3.0.3
djangorestframework==3.11.0
mysqlclient==1.4.6
pays==0.2.3
Pillow==7.0.0
PyMySQL==0.9.3
pytz==2019.3
sqlparse==0.3.0
whitenoise==5.0.1
